requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
dj-database-url==0.5.0
Django==4.0.1
gunicorn==20.1.0
idna==3.3
psycopg2-binary==2.9.3
pycoin==0.91.20210515
python-bitcoinrpc==1.0
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==5.3.0
